server/requirements.txt
aiohttp==3.9.3
aiohttp-cors==0.7.0
aioice==0.9.0
aiortc==1.7.0
aiosignal==1.3.1
attrs==23.2.0
av==11.0.0
boto3==1.34.32
botocore==1.34.32
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.2.0
cryptography==42.0.2
cycler==0.12.1
dnspython==2.5.0
filelock==3.13.1
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.42
google-crc32c==1.5.0
gunicorn==21.2.0
idna==3.6
ifaddr==0.2.0
Jinja2==3.1.3
jmespath==1.0.1
kiwisolver==1.4.5
MarkupSafe==2.1.4
matplotlib==3.8.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
numpy==1.26.3
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.1
pillow==10.2.0
psutil==5.9.8
py-cpuinfo==9.0.0
pycairo==1.25.1
pycparser==2.21
pyee==11.1.0
pylibsrtp==0.10.0
pyOpenSSL==24.0.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.10.0
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
sympy==1.12
thop==0.1.1.post2209072238
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
ultralytics==8.1.24
urllib3==2.0.7
yarl==1.9.4
